add customer for just to keep track who is buying and to print the name on the receipt.
Customer is not full functional.
it is just implemented to the extend where we can track who is buying the order, and print the name of customer on the receipt if we want.
more on customer will be added after adding the functionality of packages and improving order management table.